backend-render/requirements.txt
fastapi==0.110.0
pydantic==2.6.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.1.0
cryptography==41.0.7
reportlab==4.0.7
openpyxl==3.1.2
pandas==2.1.4
alembic==1.8.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
